refactor : 카카오 페이지 컴포넌트화
